fix(android): bump Gradle wrapper version for 0.75
